re organizing files, got list of teams to properly display, still working on team links working properly
